+++ src/main/java/com/avp/common/item/PressureSuitArmorItem.java
@@ -19,14 +19,14 @@ public class PressureSuitArmorItem extends ArmorItem {
     }
 
     @Override
-    public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> list, TooltipFlag tooltipFlag) {
-        super.appendHoverText(stack, context, list, tooltipFlag);
+    public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
 
-        list.add(
+        tooltipComponents.add(
             Component.translatable("tooltip.avp.pressure_suit.full_set_bonus")
                 .withStyle(ChatFormatting.YELLOW)
         );
-        list.add(
+        tooltipComponents.add(
             Component.translatable("tooltip.avp.pressure_suit.water_breathing")
                 .withStyle(ChatFormatting.GREEN)
         );