+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - N/A
 
 ## ✨ What's New
+- Added tooltip for ammo chest item.
 - Added tooltips for armor items that have full set bonuses.
   - Full set buffs are colored green.
   - Full set debuffs are colored red.